refactored round trip tests into trait
